requirements-dev.txt
pyserial
pytest
pytest-html
pytest-csv
